Create documentation PDF with github actions

+++ documentation/create_pdf.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
 echo "Removing previous PDF"
-rm proteus_doc.pdf
+if [ -f proteus_doc.pdf ]; then
+  rm proteus_doc.pdf
+fi
 
 echo "Creating new PDF"
 fop -xml proteus_doc.xml -xsl xsl/proteus_pdf.xsl -pdf proteus_doc.pdf
